add user management

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -11,14 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('proj1_user')) {
+            return;
+        }
+
         Schema::create('proj1_user', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->rememberToken();
-            $table->timestamps();
+            $table->string('user_login', 1000)->primary();
+            $table->string('password', 1000)->nullable();
+            $table->string('employee_id', 1000)->nullable();
+            $table->string('employee_name', 1000)->nullable();
+            $table->string('company_code', 1000)->nullable();
+            $table->string('position_code', 1000)->nullable();
+            $table->date('date_create')->nullable();
+            $table->string('email', 300)->nullable();
+            $table->string('remember_token', 1000)->nullable();
+            $table->string('password_reset_tokens', 200)->nullable();
+            $table->string('role', 100)->nullable();
+            $table->boolean('is_active')->default(true);
         });
     }
 
@@ -27,6 +36,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (! Schema::hasTable('proj1_user')) {
+            return;
+        }
+
         Schema::dropIfExists('proj1_user');
     }
 };
